requirements.txt
feedparser>=6.0.11
python-dotenv>=1.0.0
jieba>=0.42.1
yfinance>=0.2.40
